Move formatDatePeriod.js tests to their own file
