+++ root/static/scripts/tests/index-web.js
@@ -25,6 +25,7 @@ require('./utility');
 require('./utility/age');
 require('./utility/dates');
 require('./utility/formatDate');
+require('./utility/formatDatePeriod');
 require('./utility/formatTrackLength');
 require('./utility/parseDate');
 require('./utility/sanitizedEditor');
